% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myAdd.R
\name{myAdd}
\alias{myAdd}
\title{My Add function}
\usage{
myAdd(a, b)
}
\arguments{
\item{a}{first summand}

\item{b}{second summan}
}
\value{
the sum will be returned
}
\description{
My Add function
}
\examples{
myAdd(2,3)
}
